requirements.txt
livekit-agents[openai,turn-detector]>=1.0.0
livekit-api>=0.5.0
livekit>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
websockets>=10.3
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
deepgram-sdk>=2.11.0
